Project restructuring Remove useless ISizesManager Remove useless IColursManager Remove useless IProductsManager Remove useless IProductsImagesManager Remove useless ICategoriesManager Remove useless IBrandsManager interface
Although it is a good practice to depend on abstraction rather than on concrete class, in this case this does not make sense. This is because these interfaces highly depend on the concrete class itself, thus they are not introducing any level of abstraction at all. Instead, I am just maintaining two files at once with no gain.