Asynchronous methods everywhere
Performer<ProductsController> is a service that is used to perform action by wrapping it in another action that handles exceptions. Whenever an exception in the main method occurs, it calls fallback method that can be used to provide different information to the user. It is worth to be noted that information from UserInteractionException can be shown to the user (onUserError). On the other hand, every other exception is logged but it won't be shown to the user. I should have used action filters instead of this atrocious idea.